% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaFilter.R
\name{adaFilter}
\alias{adaFilter}
\title{The main function for adaptive filtering procedures}
\usage{
adaFilter(p.matrix, r, type.I.err = c("FDR", "FWER", "PFER"),
  alpha = 0.05, fast = F)
}
\arguments{
\item{p.matrix}{an \code{M * n} matrix of p-values. \code{M} is the
number of PC hypotheses and \code{n} is the number of studies. Missing values are allowed.}

\item{r}{the required replicability level, i.e. the smallest number of nonnull individual hypotheses
for a nonnull PC hypotheses}

\item{type.I.err}{one of "FDR", "FWER" or "PFER", with the default type I error "FDR". For FWER and PFER control, the adaFilter Bonferroni procedure is performed;
for FDR control, the adaFilter BH procedure is performed.}

\item{alpha}{control level of the chosen simultaneous error. Default is \code{0.05}.}

\item{fast}{logical value, default is FALSE. If \code{fast} is TRUE, then only results for hypotheses whose filter p-values are less than \code{alpha} are computed, and for other hypotheses the results are NA. This can speed up computation when \code{M} is super large.}
}
\value{
a data frame of size \code{M * 5} where the \code{5} columns are:
\describe{
\item{decision}{a length M vector of 1 and 0s indicating whether the corresponding PC hypothesis is rejected or not. For a hypothesis, its \code{decision} element is 1 if its \code{adjusted.p} element is not greater than \code{alpha}.}
\item{adjusted.p}{the adaFilter adjusted p-values for the selection p-values. If \code{type.I.error} is "FDR", then return the adaFilter BH adjusted p-values, otherwise, return the adaFilter Bonferroni adjusted p-values.}
\item{selection.p}{the length M vector of selection p-values}
\item{filter.p}{the length M vector of filtering p-values}
\item{adj.number}{The adaFilter adjustment number. For more details, see reference.}
}
}
\description{
The main function for adaptive filtering procedures
}
\examples{
## Simulate a p-value matrix of size 1000 * 2 with 1000 partial conjunction hypotheses and 2 studies
data <- GenPMat()

## set r = 2, which means that the partial conjunction hypothesis is nonnull 
## only when in both studies the corresponding hypothesis is nonnull
## Control for FDR at level 0.1
result <- adaFilter(data$pvalue.mat, 2, "FDR", 0.1)
## print the true positives
print(which(result$decision == 1 & data$truth.pc == 1))
## plot the histogram of the selection p-values (which should be conservative)
hist(result$selection.p, breaks = 30,
main = "Histogram of the selection p-values")

}
\references{
{
J. Wang, L. Gui, W. J. Su, C. Sabatti and A. B. Owen (2020). Detecting Multiple Replicating 
Signals using Adaptive Filtering Procedures
}
}
