% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_filter.R
\docType{package}
\name{adaFilter_package}
\alias{adaFilter_package}
\alias{adaFilter_package-package}
\title{Screening for partial conjunction hypotheses controlling
using the adaptive filtering procedures}
\description{
Provides the adaptive filtering multiple testing procedures
for partial conjunction hypotheses. The procedures can more efficiently reject non-null
partial conjunction hypotheses than previously methods
especially when the number of
candidate hypotheses is large while most of them are complete nulls
}
