% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_filter.R
\name{AdaFilteringPC}
\alias{AdaFilteringPC}
\title{The main function for adaptive filtering procedures}
\usage{
AdaFilteringPC(p.matrix, alpha, r, type.I.err = c("FWER", "FDR", "PFER"))
}
\arguments{
\item{p.matrix}{an M*n matrix of p-values. M is the
number of PC hypotheses and n is the number of studies}

\item{alpha}{control level of the chosen simultaneous error}

\item{r}{the smallest number of nonnull individual hypotheses
for a nonnull partial conjunction hypotheses}

\item{type.I.err}{for FWER and PFER control,
the adaptive filtering Bonferroni procedure is used;
for FDR control, the adaptive filtering BH procedure is used}
}
\value{
a list of objects
\describe{
\item{decision}{a length M vector of 1 and 0s indicating whether the corresponding PC hypothesis is rejected or not}
\item{select.p}{the length M vector of selection p-values}
\item{filter.p}{the length M vector of filtering p-values}
}
}
\description{
The main function for adaptive filtering procedures
}
\examples{
## Simulate a p-value matrix of size 1000 * 2 with 1000 partial conjunction hypotheses and 2 studies
data <- GenPMat()

## set r = 2, which means that the partial conjunction hypothesis is nonnull 
## only when in both studies the corresponding hypothesis is nonnull
## Control for FDR at level 0.1
result <- AdaFilteringPC(data$pvalue.mat, 0.1, 2, "FDR")
## print the true positives
print(which(result$decision == 1 & data$truth.pc == 1))
## plot the histogram of the selection p-values (which should be conservative)
hist(result$select.p, breaks = 30,
main = "Histogram of the Selection P-values")

}
\references{
{
J. Wang, C. Sabatti and A. B. Owen (2016). Adaptive Filtering
Multiple Testing Procedures for Partial Conjunction
Hypotheses.
}
}

