% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classical_methods.R
\name{ClassicalMTP}
\alias{ClassicalMTP}
\title{Classical procedures for partial conjunction hypotheses}
\usage{
ClassicalMTP(p.matrix, alpha, r, method = c("Bonferroni", "Fisher", "Simes"),
  type.I.err = c("FWER", "FDR", "PFER"))
}
\arguments{
\item{p.matrix}{an M*n matrix of p-values. M is the
number of PC hypotheses and n is the number of studies}

\item{alpha}{control level of the chosen simultaneous error}

\item{r}{the smallest number of nonnull individual hypotheses
for a nonnull partial conjunction hypotheses}

\item{method}{the method to construct partial conjunction p-values. 
There are currently 3 options using either Bonferroni or Fisher or 
Simes derived partial conjunction p-value}

\item{type.I.err}{for FWER and PFER control,
the adaptive filtering Bonferroni procedure is used;
for FDR control, the adaptive filtering BH procedure is used}
}
\value{
a list of objects
\describe{
\item{decision}{a length M vector of 1 and 0s indicating whether the corresponding PC hypothesis is rejected or not}
\item{BHPC.p}{the length M vector of partial conjunction p-values}
}
}
\description{
Classical procedures for partial conjunction hypotheses
}
