% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classical_methods.R
\name{ComputeBHPC}
\alias{ComputeBHPC}
\title{Compute partial conjunction p-values}
\usage{
ComputeBHPC(p.vec, r, method = c("Bonferroni", "Fisher", "Simes"))
}
\arguments{
\item{p.vec}{a vector of p-values of length n}

\item{r}{the smallest number of nonnull individual hypotheses
for a nonnull partial conjunction hypotheses}

\item{method}{the method to construct partial conjunction p-values. 
There are currently 3 options using either Bonferroni or Fisher or 
Simes derived partial conjunction p-value}
}
\value{
the BHPC p-value
}
\description{
Compute partial conjunction p-values
}

