% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classical_methods.R
\name{computePC}
\alias{computePC}
\title{Compute partial conjunction p-values}
\usage{
computePC(p.vec, r, method = c("Bonferroni", "Fisher", "Simes"))
}
\arguments{
\item{p.vec}{a vector of p-values of length \code{n}}

\item{r}{the required replicability level, i.e. the smallest number of nonnull individual hypotheses
for a nonnull PC hypotheses}

\item{method}{the method to construct partial conjunction p-values. 
There are currently 3 options:  "Bonferroni",  "Fisher" and  
"Simes". The default is "Bonferroni".}
}
\value{
a vector of partial conjunction p-values
}
\description{
Compute partial conjunction p-values
}
