% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_filter.R
\name{FindEqualIdx}
\alias{FindEqualIdx}
\title{v1 is non-decreasing, v2 is non-increasing, find the intersection point}
\usage{
FindEqualIdx(v1, v2, end, start = 1)
}
\description{
v1 is non-decreasing, v2 is non-increasing, find the intersection point
}
\keyword{internal}
