% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_function.R
\name{GenPMat}
\alias{GenPMat}
\title{Generate p-value matrix for simulation}
\usage{
GenPMat(M = 1000, n = 2, r = 2, all.zero.frac = 0.5,
  alternative.frac = 0.1, mu = c(3, 4, 6), rho = 0, block.size = 10,
  one.sided = F)
}
\arguments{
\item{M}{number of partial conjunction hypotheses}

\item{n}{number of studies}

\item{r}{the required minimum number of nonnull studies 
for a nonnull partial conjunction hypothesis}

\item{all.zero.frac}{the fraction of complete nulls}

\item{alternative.frac}{the total fraction nonnull partial conjunction
hypotheses}

\item{mu}{a vector of possible values of effect size of nonnull individual 
hypotheses}

\item{rho}{correlation within block of screening tests}

\item{block.size}{the block size of the block dependency structure}

\item{one.sided}{either construct one-sided p-value or not}
}
\value{
a list of objects
\describe{
\item{truth.pc}{a length M vector of 1 and 0s indicating whether the 
corresponding hypothesis is a true nonnull partial conjunction hypothesis 
or not}
\item{pvalue.mat}{the individual p-value matrix of size M * n}
\item{zmat}{The corresponding z-value matrix of size M * n}
}
}
\description{
Generate p-value matrix for simulation
}
\references{
{
J. Wang, C. Sabatti and A. B. Owen (2016). Adaptive Filtering
Multiple Testing Procedures for Partial Conjunction
Hypotheses.
}
}

