% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classical_methods.R
\name{ClassicalMTP}
\alias{ClassicalMTP}
\title{Classical procedures for partial conjunction hypotheses}
\usage{
ClassicalMTP(p.matrix, alpha, r, method = c("Bonferroni", "Fisher",
  "Simes"), type.I.err = c("FWER", "FDR", "PFER"))
}
\arguments{
\item{p.matrix}{an \code{M * n} matrix of p-values. \code{M} is the
number of PC hypotheses and \code{n} is the number of studies. Missing values are allowed.}

\item{alpha}{control level of the chosen simultaneous error. Default is \code{0.05}.}

\item{r}{the required replicability level, i.e. the smallest number of nonnull individual hypotheses
for a nonnull PC hypotheses}

\item{method}{the method to construct partial conjunction p-values. 
There are currently 3 options:  "Bonferroni",  "Fisher" and  
"Simes". The default is "Bonferroni".}

\item{type.I.err}{one of "FDR", "FWER" or "PFER", with the default type I error "FDR". For FWER and PFER control, the adaFilter Bonferroni procedure is performed;
for FDR control, the adaFilter BH procedure is performed.}
}
\value{
a list of objects
\describe{
\item{decision}{a length M vector of 1 and 0s indicating whether the corresponding PC hypothesis is rejected or not}
\item{PC.pvalues}{the length M vector of partial conjunction p-values}
}
}
\description{
Classical procedures for partial conjunction hypotheses
}
